// src/generated/models/access-control-entry-model.ts
import type { AccessRole } from './access-role-model';
import type { Application } from './application-model';
import type { Environment } from './environment-model';
import type { RequestedUser } from './requested-user-model';

export interface AccessControlEntry {
  /**
   * @displayName Access Control Entry
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Access Control Entry Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  accessControlEntryName: string;
  /**
   * @displayName (Deprecated) Access Role
   */
  accessRole?: Pick<AccessRole, 'id' | 'accessRoleName'>;
  /**
   * @displayName Active State
   * @validationRule Required for create/update operations
   */
  activeState: boolean;
  /**
   * @displayName Application
   * @validationRule Required for create/update operations
   */
  application: Pick<Application, 'id' | 'applicationName'>;
  /**
   * @displayName Date Provisioned
   */
  dateProvisioned?: string;
  /**
   * @displayName (Deprecated) Environment
   */
  environment?: Pick<Environment, 'id' | 'environmentName'>;
  /**
   * @displayName Granted By
   */
  grantedBy?: string;
  /**
   * @displayName Granted Date
   */
  grantedDate?: string;
  /**
   * @displayName Requested User
   * @validationRule Required for create/update operations
   */
  requestedUser: Pick<RequestedUser, 'id' | 'fullName'>;
}

export const _AccessControlEntry = 'AccessControlEntry' as const;