// src/generated/models/requested-user-model.ts
import type { AccessRole } from './access-role-model';
import type { Application } from './application-model';
import type { Environment } from './environment-model';
import type { ProvisioningRequest } from './provisioning-request-model';

export const RequestedUserStatusKeyToLabel = {
  'New': 'New',
  'SentToProvisioning': 'Sent to Provisioning',
  'Provisioned': 'Provisioned',
  'GCCSyncComplete': 'GCC Sync Complete',
  'AccessConfirmed': 'Access Confirmed',
  'NotificationReady': 'Notification Ready',
  'Notified': 'Notified',
  'Exception': 'Exception',
  'Cancelled': 'Cancelled'
} as const;
export type RequestedUserStatusKey = keyof typeof RequestedUserStatusKeyToLabel;

export interface RequestedUser {
  /**
   * @displayName Requested User
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Full Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  fullName: string;
  /**
   * @displayName (Deprecated) Access Role
   */
  accessRole?: Pick<AccessRole, 'id' | 'accessRoleName'>;
  /**
   * @displayName Active State
   */
  activeState?: boolean;
  /**
   * @displayName Application
   * @validationRule Required for create/update operations
   */
  application: Pick<Application, 'id' | 'applicationName'>;
  /**
   * @displayName BSC ID
   * @validationRule Required for create/update operations
   */
  bSCID: string;
  /**
   * @displayName Email
   * @validationRule Required for create/update operations
   */
  email: string;
  /**
   * @displayName (Deprecated) Environment
   */
  environment?: Pick<Environment, 'id' | 'environmentName'>;
  /**
   * @displayName (Deprecated) Exception Reason
   */
  exceptionReason?: string;
  /**
   * @displayName Normalized Email
   * @validationRule Required for create/update operations
   */
  normalizedEmail: string;
  /**
   * @displayName (Deprecated) Notes
   */
  notes?: string;
  /**
   * @displayName (Deprecated) Provisioning Request
   */
  provisioningRequest?: Pick<ProvisioningRequest, 'id' | 'provisioningRequestName'>;
  /**
   * @displayName (Deprecated) Provisioning Ticket Reference
   */
  provisioningTicketReference?: string;
  /**
   * @displayName (Deprecated) Status
   * @validationRule Required for create/update operations
   */
  statusKey: RequestedUserStatusKey;
}

export const _RequestedUser = 'RequestedUser' as const;